+++ defaults/controllers/_default/resourceJoin/index.js
@@ -13,14 +13,10 @@ exports = module.exports = {
 	// functions
 
 	init: function(defaultFn) {
-		return resourceIndexAction.init.call(this, defaultFn).bind(this)
-		.then(function() {
-			// sort by target's name
-			this.sortDefault = this.route.targetModel.name + 'Id';
+		// sort by target's name
+		this.sortDefault = this.route.targetModel.name + 'Id';
 
-			// remove fields for parent name
-			delete this.fields[this.route.parent.model.name + 'Id'];
-		});
+		return resourceIndexAction.init.call(this, defaultFn);
 	},
 
 	start: resourceIndexAction.start,